+++ src/AppserverIo/Appserver/DependencyInjectionContainer/ObjectManagerFactory.php
@@ -22,7 +22,7 @@ namespace AppserverIo\Appserver\DependencyInjectionContainer;
 
 use AppserverIo\Storage\StackableStorage;
 use AppserverIo\Psr\Application\ApplicationInterface;
-use AppserverIo\Appserver\Application\Interfaces\ManagerConfigurationInterface;
+use AppserverIo\Appserver\Core\Api\Node\ManagerNodeInterface;
 
 /**
  * The factory for the object manager.
@@ -39,12 +39,12 @@ class ObjectManagerFactory
     /**
      * The main method that creates new instances in a separate context.
      *
-     * @param \AppserverIo\Psr\Application\ApplicationInterface                           $application          The application instance to register the class loader with
-     * @param \AppserverIo\Appserver\Application\Interfaces\ManagerConfigurationInterface $managerConfiguration The manager configuration
+     * @param \AppserverIo\Psr\Application\ApplicationInterface         $application          The application instance to register the class loader with
+     * @param \AppserverIo\Appserver\Core\Api\Node\ManagerNodeInterface $managerConfiguration The manager configuration
      *
      * @return void
      */
-    public static function visit(ApplicationInterface $application, ManagerConfigurationInterface $managerConfiguration)
+    public static function visit(ApplicationInterface $application, ManagerNodeInterface $managerConfiguration)
     {
 
         // load the configured descriptors from the configuration